requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
Django==4.2.5
django-extensions==3.2.3
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
h11==0.14.0
idna==3.4
inflection==0.5.1
Markdown==3.4.4
mysqlclient==2.2.0
outcome==1.2.0
packaging==23.2
pycparser==2.21
PyMySQL==1.1.0
PySocks==1.7.1
python-dotenv==1.0.0
pytube==15.0.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.13.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.0.6
webdriver-manager==4.0.1
wsproto==1.2.0
